Only remove images if there was a new one committed
Before, it looks like it was removing images
but sometimes not committing a new image.
Especially at the end of the build,
when it's really important that there be
1 image available.
Sometimes there was no image,
so it couldn't open the finalTerminal.
Also, don't clean up the images
on the timeout.
There's already a function that does that
when closing all terminals.